feat: Add OS and Kubelet disk type options.
OS Disk type can be either Managed or Temporary.
Kubelet Disk type can be either OS or Ephemeral.
The default behaviour for the option, and when not set, is to use
Managed and OS for the OS and Kubelet respectively. When set to
Temporary and Ephemeral the local temporary storage will be used if
available on the node.

Signed-off-by: Carus Kyle <caruskyle@gmail.com>